[Chore] #36 divided firebaseConnector in 3 collections
